// theme/colors.generated.ts
export const Colors = {
  "scheme": {
    "light": {
      "primary": "#8C4A60",
      "surfaceTint": "#8C4A60",
      "onPrimary": "#FFFFFF",
      "primaryContainer": "#FFD9E2",
      "onPrimaryContainer": "#703348",
      "secondary": "#74565F",
      "onSecondary": "#FFFFFF",
      "secondaryContainer": "#FFD9E2",
      "onSecondaryContainer": "#5A3F47",
      "tertiary": "#7C5635",
      "onTertiary": "#FFFFFF",
      "tertiaryContainer": "#FFDCC1",
      "onTertiaryContainer": "#613F20",
      "error": "#BA1A1A",
      "onError": "#FFFFFF",
      "errorContainer": "#FFDAD6",
      "onErrorContainer": "#93000A",
      "background": "#FFF8F8",
      "onBackground": "#22191C",
      "surface": "#FFF8F8",
      "onSurface": "#22191C",
      "surfaceVariant": "#F2DDE1",
      "onSurfaceVariant": "#514347",
      "outline": "#837377",
      "outlineVariant": "#D5C2C6",
      "shadow": "#000000",
      "scrim": "#000000",
      "inverseSurface": "#372E30",
      "inverseOnSurface": "#FDEDEF",
      "inversePrimary": "#FFB1C8",
      "primaryFixed": "#FFD9E2",
      "onPrimaryFixed": "#3A071D",
      "primaryFixedDim": "#FFB1C8",
      "onPrimaryFixedVariant": "#703348",
      "secondaryFixed": "#FFD9E2",
      "onSecondaryFixed": "#2B151C",
      "secondaryFixedDim": "#E3BDC6",
      "onSecondaryFixedVariant": "#5A3F47",
      "tertiaryFixed": "#FFDCC1",
      "onTertiaryFixed": "#2E1500",
      "tertiaryFixedDim": "#EFBD94",
      "onTertiaryFixedVariant": "#613F20",
      "surfaceDim": "#E6D6D9",
      "surfaceBright": "#FFF8F8",
      "surfaceContainerLowest": "#FFFFFF",
      "surfaceContainerLow": "#FFF0F2",
      "surfaceContainer": "#FAEAED",
      "surfaceContainerHigh": "#F5E4E7",
      "surfaceContainerHighest": "#EFDFE1"
    },
    "dark": {
      "primary": "#FFB1C8",
      "surfaceTint": "#FFB1C8",
      "onPrimary": "#541D32",
      "primaryContainer": "#703348",
      "onPrimaryContainer": "#FFD9E2",
      "secondary": "#E3BDC6",
      "onSecondary": "#422931",
      "secondaryContainer": "#5A3F47",
      "onSecondaryContainer": "#FFD9E2",
      "tertiary": "#EFBD94",
      "onTertiary": "#48290B",
      "tertiaryContainer": "#613F20",
      "onTertiaryContainer": "#FFDCC1",
      "error": "#FFB4AB",
      "onError": "#690005",
      "errorContainer": "#93000A",
      "onErrorContainer": "#FFDAD6",
      "background": "#191113",
      "onBackground": "#EFDFE1",
      "surface": "#191113",
      "onSurface": "#EFDFE1",
      "surfaceVariant": "#514347",
      "onSurfaceVariant": "#D5C2C6",
      "outline": "#9E8C90",
      "outlineVariant": "#514347",
      "shadow": "#000000",
      "scrim": "#000000",
      "inverseSurface": "#EFDFE1",
      "inverseOnSurface": "#372E30",
      "inversePrimary": "#8C4A60",
      "primaryFixed": "#FFD9E2",
      "onPrimaryFixed": "#3A071D",
      "primaryFixedDim": "#FFB1C8",
      "onPrimaryFixedVariant": "#703348",
      "secondaryFixed": "#FFD9E2",
      "onSecondaryFixed": "#2B151C",
      "secondaryFixedDim": "#E3BDC6",
      "onSecondaryFixedVariant": "#5A3F47",
      "tertiaryFixed": "#FFDCC1",
      "onTertiaryFixed": "#2E1500",
      "tertiaryFixedDim": "#EFBD94",
      "onTertiaryFixedVariant": "#613F20",
      "surfaceDim": "#191113",
      "surfaceBright": "#413739",
      "surfaceContainerLowest": "#140C0E",
      "surfaceContainerLow": "#22191C",
      "surfaceContainer": "#261D20",
      "surfaceContainerHigh": "#31282A",
      "surfaceContainerHighest": "#3C3235"
    }
  }
};